ARPA-359 Refactoring
