+++ Orso.Arpa.Application/AuthApplication/SetRoleAuthorizationHandler.cs
@@ -8,14 +8,13 @@ using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
-using Orso.Arpa.Application.AuthApplication;
 using Orso.Arpa.Domain.Entities;
 using Orso.Arpa.Domain.Identity;
 using Orso.Arpa.Domain.Interfaces;
 using Orso.Arpa.Domain.Roles;
 using Orso.Arpa.Infrastructure.Authorization.AuthorizationRequirements;
 
-namespace Orso.Arpa.Infrastructure.Authorization.AuthorizationHandlers
+namespace Orso.Arpa.Application.AuthApplication
 {
     public class SetRoleAuthorizationHandler : AuthorizationHandler<SetRoleAuthorizationRequirement>
     {
