+++ Orso.Arpa.Application/SelectValueApplication/SelectValueDto.cs
@@ -1,12 +1,15 @@
 using AutoMapper;
 using Orso.Arpa.Application.General;
 using Orso.Arpa.Domain.Entities;
+using Orso.Arpa.Infrastructure.Localization;
 
 namespace Orso.Arpa.Application.SelectValueApplication
 {
     public class SelectValueDto : BaseEntityDto
     {
+        [Translate(nameof(SelectValueDto))]
         public string Name { get; set; }
+        [Translate(nameof(SelectValueDto))]
         public string Description { get; set; }
     }
